+++ Hazel/src/Platform/Windows/WindowsWindow.h
@@ -18,7 +18,7 @@ namespace Hazel {
 
 		inline unsigned int GetWidth() const override { return m_Data.Width; }
 		inline unsigned int GetHeight() const override { return m_Data.Height; }
-		inline float GetAspectRatio() const override { return m_Data.Width / m_Data.Height; }
+		inline float GetAspectRatio() const override { return (float)m_Data.Width / (float)m_Data.Height; }
 
 		//Window attributes
 		inline void SetEventCallback(const EventCallbackFn& callback) override { m_Data.EventCallback = callback; }
